requirements.txt
torch==2.1.1
torchvision==0.16.1
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
matplotlib==3.8.2
segment-anything==1.0
